requirements.txt
aiofiles==23.2.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
boto3==1.34.128
botocore==1.34.128
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cohere==5.9.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.114.2
fastavro==1.9.7
ffmpy==0.4.0
filelock==3.16.0
fonttools==4.53.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
#gradio==4.44.0
gradio==3.35
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.7
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
#markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.8.1
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.3
prompt_toolkit==3.0.47
protobuf==5.28.1
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==2.9.1
pydantic_core==2.23.3
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.2
#pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
ruff==0.6.5
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==3.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.38.5
streamlit==1.38.0
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
tomlkit==0.12.0
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typer==0.12.5
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
uvicorn==0.30.6
watchdog==4.0.2
wcwidth==0.2.13
websockets==12.0
zipp==3.20.2
